backend/requirements-dev.txt
# Dépendances de développement supplémentaires
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==22.0.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3
mypy==1.7.1
django-stubs==4.2.6
ipython==8.18.1
django-debug-toolbar==4.2.0
